requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6

# AI/ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.26.0
opencv-python>=4.9.0

# Testing
pytest>=8.0.0
httpx>=0.26.0
pillow>=10.2.0

# Development
ruff>=0.2.0
